enrichment workers: remove the duplicated fetch_alerts method for a generic one
